refactor(api): move the connection-token exchange out of index.ts
The Express handler, its error-status map, and the header flattening now
live in src/routes/connection-token.ts; index.ts just registers the
route. The shared unknown-error-to-message step moves to api-credential,
next to the error names it reads.
